Changed some more things.
