+++ tsconfig.json
@@ -6,9 +6,8 @@
     "moduleResolution": "node",         
     "lib": ["es7", "dom"],              
     "declaration": false,               
-    "sourceMap": false,                 
-	"removeComments": true,
-	"inlineSourceMap": true,             
+    "sourceMap": true,                 
+	"removeComments": true,             
     "isolatedModules": false,           
     "strict": true,                    
     "noImplicitAny": true,
